Pick symbols based on best-matching luminance value
Add a tool to define SymbolSets with luminance values. This can be found by visiting `/icons`.

Previously, the logic that picked which symbol to use for a pixel assumed that all the symbols where uniformly distributed by their perceived "brightness" values. So if a pixel had a lumninance value of 177 (50%), and if the symbol set had 10 symbols, it would pick the 5th symbol. Unfortunately, this does not reflect the reality of symbol sets.