Only use filled symbols for some recipes
A bug in the web tool made it so that only filled symbols were used (a boolean was written as a string, so it always evaluated as true). The designers, however, liked the result. Instead of reintroducing the bug, we are going to force all the symbols to be filled (for the relevant recipes/symbol sets) to recreate the effect.